requirements.txt
requests
discord.py
python-dotenv
openai
google-generativeai
aiohttp
